requirements.txt
Flask==3.0.0
Flask-Cors==5.0.0
Werkzeug==3.1.3
gunicorn==25.3.0
numpy==1.26.4
Pillow==10.2.0
tensorflow-cpu==2.19.0